Module subspace_cli::utils

Constants

KEEP_LAST_N_FILE ğŸ”’

for how long a log file should be valid

Traits

IntoEyre

IntoEyreFuture

IntoEyreStream

Functions

apply_extra_options

cache_directory_getter ğŸ”’

generates a cache path from the given path

custom_log_dir ğŸ”’

returns OS specific log directory

data_dir_getter ğŸ”’

directory_parser ğŸ”’

the provided path should be an existing directory

get_user_input ğŸ”’

gets the input from the user for a given prompt

install_tracing ğŸ”’

install a logger for the application

into_eyre_err ğŸ”’

is_default

node_directory_getter ğŸ”’

generates a node path from the given path

node_name_parser ğŸ”’

node name should be ascii, and should begin/end with whitespace

open_log_dir ğŸ”’

opens log directory

plot_directory_getter ğŸ”’

generates a plot path from the given path

print_ascii_art ğŸ”’

<3

print_run_executable_command ğŸ”’

print_version ğŸ”’

prints the version of the crate

provider_storage_dir_getter ğŸ”’

raise_fd_limit ğŸ”’

reward_address_parser ğŸ”’

check for a valid SS58 address

size_parser ğŸ”’

utilize ByteSize crate for the validation

spawn_task ğŸ”’

support_message ğŸ”’

in case of any error, display this message

yes_or_no_parser ğŸ”’