Module subspace_cli::utils

Constants

KEEP_LAST_N_FILE ğŸ”’

for how long a log file should be valid

Functions

cache_directory_getter ğŸ”’

generates a cache path from the given path

custom_log_dir ğŸ”’

returns OS specific log directory

data_dir_getter ğŸ”’

get_user_input ğŸ”’

gets the input from the user for a given prompt

install_tracing ğŸ”’

install a logger for the application

node_directory_getter ğŸ”’

generates a node path from the given path

node_name_parser ğŸ”’

node name should be ascii, and should begin/end with whitespace

plot_directory_getter ğŸ”’

generates a plot path from the given path

plot_directory_parser ğŸ”’

the provided path should be an existing directory

print_ascii_art ğŸ”’

<3

print_version ğŸ”’

prints the version of the crate

raise_fd_limit ğŸ”’

reward_address_parser ğŸ”’

check for a valid SS58 address

size_parser ğŸ”’

utilize ByteSize crate for the validation

support_message ğŸ”’

in case of any error, display this message

yes_or_no_parser ğŸ”’